refactor(plugins): share install E2E helper
