+++ apps/web/e2e/tests/plugins/mobile-plugin-nav.spec.ts
@@ -13,6 +13,7 @@ import path from "node:path";
 import type { Browser, BrowserContext, Page } from "@playwright/test";
 import { expect, test } from "../../fixtures/test-base";
 import { PrAssetCapture } from "../../helpers/pr-asset-capture";
+import { holdPluginInstallResponse } from "../../helpers/plugin-install";
 
 const PLUGIN_ID = "kandev-plugin-e2e";
 const NAV_ITEM_ID = "e2e-hello";
@@ -21,39 +22,6 @@ const PACKAGE_PATH = path.resolve(
   __dirname,
   "../../../../../apps/backend/.build/kandev-plugin-e2e-1.0.0.tar.gz",
 );
-
-async function holdPluginInstallResponse(page: Page) {
-  let release = () => {};
-  let requestStarted = false;
-  let markResponseSettled = () => {};
-  const responseHeld = new Promise<void>((resolve) => {
-    release = resolve;
-  });
-  const responseSettled = new Promise<void>((resolve) => {
-    markResponseSettled = resolve;
-  });
-  let markRequestSeen = () => {};
-  const requestSeen = new Promise<void>((resolve) => {
-    markRequestSeen = resolve;
-  });
-
-  await page.route("**/api/plugins/install", async (route) => {
-    requestStarted = true;
-    markRequestSeen();
-    try {
-      await responseHeld;
-      await route.fulfill({
-        status: 500,
-        contentType: "application/json",
-        body: JSON.stringify({ error: "install failed" }),
-      });
-    } finally {
-      markResponseSettled();
-    }
-  });
-
-  return { requestSeen, responseSettled, release, requestStarted: () => requestStarted };
-}
 
 function pluginExecutablePath(installPath: string): string {
   const serverDir = path.join(installPath, "server");
