+++ apps/web/e2e/tests/plugins/plugins.spec.ts
@@ -48,6 +48,7 @@ import type { Page } from "@playwright/test";
 import { expect, test } from "../../fixtures/test-base";
 import { SessionPage } from "../../pages/session-page";
 import type { ApiClient } from "../../helpers/api-client";
+import { holdPluginInstallResponse } from "../../helpers/plugin-install";
 
 const PLUGIN_ID = "kandev-plugin-e2e";
 const NAV_ITEM_ID = "e2e-hello";
@@ -82,39 +83,6 @@ async function uploadPackage(page: Page, filePath: string) {
   await page.getByTestId("install-plugin-tab-upload").click();
   await page.getByTestId("install-plugin-file-input").setInputFiles(filePath);
   await page.getByTestId("install-plugin-upload-submit").click();
-}
-
-async function holdPluginInstallResponse(page: Page) {
-  let release = () => {};
-  let requestStarted = false;
-  let markResponseSettled = () => {};
-  const responseHeld = new Promise<void>((resolve) => {
-    release = resolve;
-  });
-  const responseSettled = new Promise<void>((resolve) => {
-    markResponseSettled = resolve;
-  });
-  let markRequestSeen = () => {};
-  const requestSeen = new Promise<void>((resolve) => {
-    markRequestSeen = resolve;
-  });
-
-  await page.route("**/api/plugins/install", async (route) => {
-    requestStarted = true;
-    markRequestSeen();
-    try {
-      await responseHeld;
-      await route.fulfill({
-        status: 500,
-        contentType: "application/json",
-        body: JSON.stringify({ error: "install failed" }),
-      });
-    } finally {
-      markResponseSettled();
-    }
-  });
-
-  return { requestSeen, responseSettled, release, requestStarted: () => requestStarted };
 }
 
 async function uninstallViaApi(apiClient: ApiClient) {
